Add more use cases

+++ Babylon/src/Modules/Channels/Babylon.Modules.Channels.Application/Channels/DeleteMemberFormChannel/DeleteMemberFormChannelCommandHandler.cs
@@ -2,13 +2,29 @@
 using Babylon.Common.Domain;
 using Babylon.Modules.Channels.Application.Abstractions.Data;
 using Babylon.Modules.Channels.Domain.Channels;
+using Babylon.Modules.Channels.Domain.Members;
 
 namespace Babylon.Modules.Channels.Application.Channels.DeleteMemberFormChannel;
-internal sealed class DeleteMemberFormChannelCommandHandler(IChannelMemberRepository channelMemberRepository, IUnitOfWork unitOfWork) : ICommandHandler<DeleteMemberFormChannelCommand>
+internal sealed class DeleteMemberFormChannelCommandHandler(
+    IChannelMemberRepository channelMemberRepository, 
+    IUnitOfWork unitOfWork,
+    IChannelRepository channelRepository) : ICommandHandler<DeleteMemberFormChannelCommand>
 {
     public async Task<Result> Handle(DeleteMemberFormChannelCommand request, CancellationToken cancellationToken)
     {
-        ChannelMember channelMember = await channelMemberRepository.GetChannelMember(request.ChannelId, request.MemberId);
+        Channel? channel = await channelRepository.GetChannel(request.ChannelId);
+        if (channel is null)
+        {
+            throw new InvalidOperationException("Channel was not found");
+        }
+
+        ChannelMember? channelMember = await channelMemberRepository.GetChannelMember(request.ChannelId, request.Id);
+
+        if(channelMember == null)
+        {
+            throw new InvalidOperationException("Member was not part of the channel");
+        }
+
         await channelMemberRepository.DeleteChannelMember(channelMember);
 
         await unitOfWork.SaveChangesAsync(cancellationToken);
